requirements.txt
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.1
fastapi[standard]==0.114.2
streamlit==1.38.0
joblib==1.4.2
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
